Add comments to controller classes
The controller classes are one of the most important parts of the program. Adding comments to improve readability and maintainability.

Other Changes/Fixes:
- Added backport of std::clamp from C++17
- Improved logic for checking C++ version in util.h
- Changed MotorConstants in Mpc class to be a struct instead
- Fixed kMaxDistance field in FcYyUsV4 class not using the correct max value

+++ inc/fc_yy_us_v4.h
@@ -29,7 +29,7 @@ using libbase::k60::Pin;
 class FcYyUsV4 {
  public:
   static constexpr uint16_t kMinDistance = 0;
-  static constexpr uint16_t kMaxDistance = std::numeric_limits<unsigned int>::max();
+  static constexpr uint16_t kMaxDistance = std::numeric_limits<uint16_t>::max();
 
   /**
    * @param pin Name of Pin connected to the sensor
